Implement proper targeting of `_top` and `_parent`
When a window context targets either `_parent` or `_top`, it should
instead target itself, even if it is a child window.  Targeting `_top`
should continue checking the parent until a window context is found. For
this reason, when targeting `_top` or `_parent`, the context should be
checked to determine if it is a frame context or a window context.