Use document from nested context instead of request
The document in the request is only updated when the `src` or `srcdoc`
is updated.  Navigation elements with named targets only update the
named child `BrowsingContext` it is able to find, or creating a new one
if one is not found.  Thus, for an iframe's `CurrentDocument` to be
updateable by a navigation element, it needs to use its nested context,
rather than the request.  Also, when the iframe is being set up, a
nested context needs to be made and, if named, needs to be added to the
parent context so that it is available for navigation elements targeting
the iframe even if it does not have an intial document to load.