Add test case for iframe nested in unnamed iframe
Targeting an named iframe works even if it is nested within an unnamed
iframe.  This is being added due to the previous design of nested
browsing contexts where unnamed child browsing contexts are not added
to the parent's collection of child brosing context, which prevents
(or at least makes very difficult) to navigate from a parent browsing
context to a child browsing context when attempting to find the target
browsing context.