Make BrowsingContext.FindChild recursive
When attempting to find a child browsing context by name, if one is not
found for the current context, then all child contexts should be
searched recursively.  This allows a browsing context to be targeted
even if it is deeply nested.  In order to get all child browsing
contexts, including annonymous browsing contexts, all attached browsing
contexts on the active document is used.