Search auxilary top-level browsing contexts
After performing a recursive depth first search for a target on all
browsing contexts starting from the first and continuing until the
top-most, then all auxilary browsing context should be searched.  Also,
any new child context created which is not a frame context should be an
auxilary browsing context, and added as a direct child of the top-most
browsing context, rather than the browsing context attempting to create
it.  The order for searching auxilary browsing contexts is currently up
to the implementation to decide, so long as it is consistent.  The order
chosen in this implementation is the order the auxilary browsing
contexts were created.